refactor(security): decouple server salt helper
